src/eYcel/yaml_handler.py
"""
YAML rules file generation, validation, and loading.
"""
from typing import Any, Dict, List, Tuple
import yaml


def generate_rules(metadata: Dict[str, Any], columns: Dict[str, Dict]) -> Dict[str, Any]:
    """
    Generate rules dictionary from metadata and column configs.
    
    Args:
        metadata: File metadata (original filename, timestamp, etc.)
        columns: Column transformation configurations
        
    Returns:
        Complete rules dictionary
    """
    # TODO: Implement rules generation
    pass


def save_rules(rules_dict: Dict[str, Any], filepath: str) -> None:
    """
    Save rules to YAML file.
    
    Args:
        rules_dict: Rules dictionary
        filepath: Path to save YAML file
    """
    # TODO: Implement YAML saving
    pass


def load_rules(filepath: str) -> Dict[str, Any]:
    """
    Load and validate rules from YAML file.
    
    Args:
        filepath: Path to rules YAML file
        
    Returns:
        Rules dictionary
    """
    # TODO: Implement YAML loading with validation
    pass


def validate_rules(rules_dict: Dict[str, Any]) -> Tuple[bool, List[str]]:
    """
    Validate rules dictionary structure.
    
    Args:
        rules_dict: Rules to validate
        
    Returns:
        Tuple of (is_valid, list_of_errors)
    """
    # TODO: Implement schema validation
    pass


def sanitize_rules(rules_dict: Dict[str, Any]) -> Dict[str, Any]:
    """
    Ensure rules contain no original data values.
    
    Args:
        rules_dict: Rules to sanitize
        
    Returns:
        Sanitized rules dictionary
    """
    # TODO: Implement sanitization
    pass


def get_yaml_schema() -> Dict[str, Any]:
    """
    Return the expected YAML schema for validation.
    
    Returns:
        Schema dictionary
    """
    # TODO: Define and return schema
    pass
